requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Faker==24.11.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
